adding user based task list

+++ todolist/views.py
@@ -3,21 +3,28 @@ from django.http import HttpResponse
 from todolist.forms import TaskForm
 from todolist.models import TaskList
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
-# Create your views here.
+
 def index(request):
-    return render(request, 'index.html')
+    parameter = {
+        'title': 'TaskManager App',
+        
+    }
+    return render(request, 'index.html', parameter)
 
-
+@login_required
 def tasklist(request):
     if request.method == "POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.owner = request.user
+            instance.save()
         messages.success(request, ("new task added"))
         return redirect("tasklist")
     else:
-        all_task = TaskList.objects.all()
+        all_task = TaskList.objects.filter(owner=request.user)
         
         ## pending and completed task
         pending_task = 0
@@ -36,14 +43,17 @@ def tasklist(request):
         }
         return render(request, "task.html", parameters)
     
-    
+@login_required 
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
-    messages.success(request, ("task \"" + task.task + "\" deleted."))
+    if task.owner == request.user:
+        task.delete()
+        messages.success(request, ("task \"" + task.task + "\" deleted."))
+    else:
+        messages.success(request, ("Access Restricted, You'cant delete this task."))
     return redirect("tasklist")
 
-
+@login_required
 def edit_task(request, task_id):
     if request.method == "POST":
         task_obj = TaskList.objects.get(pk=task_id)
@@ -57,9 +67,12 @@ def edit_task(request, task_id):
         task_obj = TaskList.objects.get(pk=task_id)
         return render(request, 'edit.html', {'obj':task_obj})
     
-    
+@login_required  
 def complete_uncomplete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = not task.done
-    task.save()
+    if task.owner == request.user:
+        task.done = not task.done
+        task.save()
+    else:
+        messages.success(request, ("Access Restricted, You'cant upadte task status."))
     return redirect('tasklist')
